requirements.txt
Django==2.2.4
psycopg2-binary==2.8.3
django-environ==0.4.5
gunicorn==19.9.0
whitenoise==4.1.3
requests==2.22.0
python-dateutil==2.8.0
docker==4.0.2
beautifulsoup4==4.8.0
lxml==4.4.0
django-storages==1.7.1
boto3==1.5.0
PyGithub==1.43.8
sentry-sdk==0.10.2
vcrpy==2.0.1
django-db-geventpool==3.1.0
gevent==1.4.0
pylint==2.3.1
pylint-django==2.0.11
black==19.3b0
